+++ mslib/utils/airdata.py
@@ -6,10 +6,10 @@
 
     Functions for getting and downloading airspaces and airports.
 
-    This file is part of mss.
+    This file is part of MSS.
 
     :copyright: Copyright 2021 May Bär
-    :copyright: Copyright 2021-2022 by the mss team, see AUTHORS.
+    :copyright: Copyright 2021-2022 by the MSS team, see AUTHORS.
     :license: APACHE-2.0, see LICENSE for details.
 
     Licensed under the Apache License, Version 2.0 (the "License");
@@ -36,7 +36,7 @@ from PyQt5 import QtWidgets
 import logging
 import time
 
-from mslib.msui.constants import MSS_CONFIG_PATH
+from mslib.msui.constants import MSUI_CONFIG_PATH
 
 
 _airspaces = []
@@ -81,16 +81,18 @@ def download_progress(file_path, url, progress_callback=lambda f: logging.info(f
 
 def get_airports(force_download=False):
     """
-    Gets or downloads the airports.csv in ~/.config/mss and returns all airports within
+    Gets or downloads the airports.csv in ~/.config/msui and returns all airports within
     """
     global _airports, _airports_mtime
-    file_exists = os.path.exists(os.path.join(MSS_CONFIG_PATH, "airports.csv"))
+    file_exists = os.path.exists(os.path.join(MSUI_CONFIG_PATH, "airports.csv"))
 
-    if _airports and file_exists and os.path.getmtime(os.path.join(MSS_CONFIG_PATH, "airports.csv")) == _airports_mtime:
+    if _airports and file_exists and \
+            os.path.getmtime(os.path.join(MSUI_CONFIG_PATH, "airports.csv")) == _airports_mtime:
         return _airports
 
-    is_outdated = file_exists \
-        and (time.time() - os.path.getmtime(os.path.join(MSS_CONFIG_PATH, "airports.csv"))) > 60 * 60 * 24 * 30
+    time_outdated = 60 * 60 * 24 * 30  # 30 days
+    is_outdated = file_exists and (time.time() - os.path.getmtime(os.path.join(MSUI_CONFIG_PATH,
+                                                                               "airports.csv"))) > time_outdated
 
     if (force_download or is_outdated or not file_exists) \
             and QtWidgets.QMessageBox.question(None, "Allow download", f"You selected airports to be "
@@ -99,11 +101,11 @@ def get_airports(force_download=False):
                                                 if not force_download else "") + "\nIs now a good time?",
                                                QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No) \
             == QtWidgets.QMessageBox.Yes:
-        download_progress(os.path.join(MSS_CONFIG_PATH, "airports.csv"), "https://ourairports.com/data/airports.csv")
+        download_progress(os.path.join(MSUI_CONFIG_PATH, "airports.csv"), "https://ourairports.com/data/airports.csv")
 
-    if os.path.exists(os.path.join(MSS_CONFIG_PATH, "airports.csv")):
-        with open(os.path.join(MSS_CONFIG_PATH, "airports.csv"), "r", encoding="utf8") as file:
-            _airports_mtime = os.path.getmtime(os.path.join(MSS_CONFIG_PATH, "airports.csv"))
+    if os.path.exists(os.path.join(MSUI_CONFIG_PATH, "airports.csv")):
+        with open(os.path.join(MSUI_CONFIG_PATH, "airports.csv"), "r", encoding="utf8") as file:
+            _airports_mtime = os.path.getmtime(os.path.join(MSUI_CONFIG_PATH, "airports.csv"))
             return list(csv.DictReader(file, delimiter=","))
 
     else:
@@ -132,7 +134,7 @@ def update_airspace(force_download=False, countries=["de"]):
     """
     global _airspaces, _airspaces_mtime
     for country in countries:
-        location = os.path.join(MSS_CONFIG_PATH, f"{country}_asp.xml")
+        location = os.path.join(MSUI_CONFIG_PATH, f"{country}_asp.xml")
         url = _airspace_download_url.format(country)
         data = [airspace for airspace in get_available_airspaces() if airspace[0].startswith(country)][0]
         file_exists = os.path.exists(location)
@@ -151,18 +153,18 @@ def update_airspace(force_download=False, countries=["de"]):
 
 def get_airspaces(countries=[]):
     """
-    Gets the .xml files in ~/.config/mss and returns all airspaces within
+    Gets the .xml files in ~/.config/msui and returns all airspaces within
     """
     global _airspaces, _airspaces_mtime
     reload = False
     files = [f"{country}_asp.xml" for country in countries]
     update_airspace(countries=countries)
-    files = [file for file in files if os.path.exists(os.path.join(MSS_CONFIG_PATH, file))]
+    files = [file for file in files if os.path.exists(os.path.join(MSUI_CONFIG_PATH, file))]
 
     if _airspaces and len(files) == len(_airspaces_mtime):
         for file in files:
             if file not in _airspaces_mtime or \
-                    os.path.getmtime(os.path.join(MSS_CONFIG_PATH, file)) != _airspaces_mtime[file]:
+                    os.path.getmtime(os.path.join(MSUI_CONFIG_PATH, file)) != _airspaces_mtime[file]:
                 reload = True
                 break
         if not reload:
@@ -171,7 +173,7 @@ def get_airspaces(countries=[]):
     _airspaces_mtime = {}
     _airspaces = []
     for file in files:
-        fpath = os.path.join(MSS_CONFIG_PATH, file)
+        fpath = os.path.join(MSUI_CONFIG_PATH, file)
         airspace = etree.parse(fpath)
         airspace = airspace.find("AIRSPACES")
         for elem in airspace:
@@ -201,5 +203,5 @@ def get_airspaces(countries=[]):
             airspace_data["polygon"] = [(float(data.split(" ")[0]), float(data.split(" ")[-1]))
                                         for data in airspace_data["polygon"].split(",")]
             _airspaces.append(airspace_data)
-            _airspaces_mtime[file] = os.path.getmtime(os.path.join(MSS_CONFIG_PATH, file))
+            _airspaces_mtime[file] = os.path.getmtime(os.path.join(MSUI_CONFIG_PATH, file))
     return _airspaces