+++ tutorials/tutorial_satellitetrack.py
@@ -1,12 +1,12 @@
 """
-    mss.tutorials.tutorial_satellitetrack
+    msui.tutorials.tutorial_satellitetrack
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
     This python script generates an automatic demonstration of how to work with remote sensing tool in topview.
-    This file is part of mss.
+    This file is part of MSS.
 
     :copyright: Copyright 2021 Hrithik Kumar Verma
-    :copyright: Copyright 2021-2022 by the mss team, see AUTHORS.
+    :copyright: Copyright 2021-2022 by the MSS team, see AUTHORS.
     :license: APACHE-2.0, see LICENSE for details.
 
     Licensed under the Apache License, Version 2.0 (the "License");
@@ -31,7 +31,7 @@ from sys import platform
 from pyscreeze import ImageNotFoundException
 
 from tutorials import screenrecorder as sr
-from mslib.msui import mss_pyui
+from mslib.msui import msui
 
 
 def initial_ops():
@@ -61,11 +61,11 @@ def call_recorder():
     rec.stop_capture()
 
 
-def call_mss():
+def call_msui():
     """
     Calls the main MSS GUI window since operations are to be performed on it only.
     """
-    mss_pyui.main()
+    msui.main()
 
 
 def automate_rs():
@@ -236,7 +236,7 @@ def main():
     This function runs the above functions as different processes at the same time and can be
     controlled from here. (This is the main process.)
     """
-    p1 = multiprocessing.Process(target=call_mss)
+    p1 = multiprocessing.Process(target=call_msui)
     p2 = multiprocessing.Process(target=automate_rs)
     p3 = multiprocessing.Process(target=call_recorder)